[FIX] sale_channel_search_engine: Prevent MRO conflicts on registry load
Avoid to create a new base model for product.product and product.category on top of the se.indexable.record model. Others addon could extends the product.product and product.category models to make it extendable. If such modules are mixed with an addon where defining a custome base class for the same models and inheriting from se.indexable.record, this will lead to an MRO conflict error when the odoo registry is initialized. This is the case for exemple when the ale_channel_search_engine_product and sale_channel_search_engine_category modules were used with the shopinvader_search_enfine module